+++ spec/support/cookbooks/resource_test/libraries/resource/python_package.rb
@@ -14,8 +14,6 @@ class Chef
 
       property :package_name, [String, Array], identity: true
       property :python, String
-      property :version, String
-      property :virtualenv, String
 
       %i[install upgrade remove].each do |act|
         action act do