+++ test/fixtures/cookbooks/test/recipes/default.rb
@@ -12,18 +12,26 @@ python_package 'pygithub' do
   python '3'
 end
 
-snu_python_package 'botocore'
-
-snu_python_package 'python_http_client' do
-  python '2'
-end
-
-snu_python_package 'python_http_client' do
+python_virtualenv '/opt/myapp' do
   python '3'
 end
 
-snu_python_virtualenv '/opt/myapp'
-
-snu_python_package 'pygithub' do
+python_package 'awscli' do
   virtualenv '/opt/myapp'
+end
+
+package 'git'
+
+directory '/usr/share/collectd'
+
+git '/usr/share/collectd/docker' do
+  repository 'https://github.com/signalfx/docker-collectd-plugin'
+end
+
+python_virtualenv '/usr/share/collectd/python' do
+  python '2'
+end
+
+pip_requirements '/usr/share/collectd/docker/requirements.txt' do
+  virtualenv '/usr/share/collectd/python'
 end