+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Snu Python Cookbook CHANGELOG
 
-This file is used to list changes made in each version of the snu_python
-cookbook.
+This file is used to list changes made in each version of the snu_python cookbook.
+
+## 1.0.0 (2018-07-11)
+
+- Delete all custom resources except snu_python
+- Remove awscli from the default Python 2 packages
 
 ## 0.2.1 (2018-08-03)